Beta 1.6 (beta build 1) (#105)
This is a partial update, planning to add more in 1.6.B2, however the present patch was quite large and decided to release.

New features and changes in this update:
 - Implemented minigame: Do or Dye (credits to Joey for the minigame itself, rewards implemented by Zera)
 - Added the missing tiles to What the Hex
 - Fixed server errors on account deletion
 - Fixed inspiration NPCs such as Konrad
 - Fix dye 9-0 error
 - Fix mistake in quest rewards
 - Fix non-unravelled enigmas showing in shop stock
 - Fix 500 error in the user details call of the API for new accounts

+++ src/main/java/org/asf/centuria/accounts/impl/FileBasedPlayerInventory.java
@@ -151,6 +151,8 @@ public class FileBasedPlayerInventory extends PlayerInventory {
 	}
 
 	private void deleteDir(File dir) {
+		if (!dir.exists())
+			return;
 		for (File subDir : dir.listFiles(t -> t.isDirectory())) {
 			deleteDir(subDir);
 		}
